requirements.txt
torch
torchvision
matplotlib
visdom
'numpy<2'
tqdm
pillow
pandas
